requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask_Cors==4.0.0
openai==1.42.0
pandas==2.2.2
pymongo==4.6.3
python-dotenv==1.0.1
PyVirtualDisplay==3.0
selenium==4.23.1
undetected_chromedriver==3.5.5
webdriver_manager==4.0.2
